fix layout of search input and add voice recognition
