+++ app/assets/stylesheets/components/_search_input.scss
@@ -57,7 +57,6 @@
 .editor-placeholder {
   background-color: white;
   text-overflow: ellipsis;
-  white-space: nowrap;
   border: none;
   overflow: auto;
   outline: none;
@@ -88,14 +87,14 @@
     max-width: 100%;
   }
 }
-.editor-footer img {
-  aspect-ratio: 2;
-  object-fit: contain;
-  object-position: center;
-  width: 80px;
-  align-self: stretch;
-  margin: auto 0;
-}
+// .editor-footer img {
+//   aspect-ratio: 2;
+//   object-fit: contain;
+//   object-position: center;
+//   width: 80px;
+//   align-self: stretch;
+//   margin: auto 0;
+// }
 .editor-button {
   border-radius: 6px;
   background-color: rgba(24, 24, 27, 1);
